Rename and clone particle-simulation html document
The 'particle-simulation .r1.html' file has been renamed to 'particle-simulation.r1.html' and cloned to a new file named 'particles.html'. This change was done to correct the previous file name that contained a misplaced space and to expand the project's documentation with a duplicated version of the 'particle-simulation.r1.html' file.